Changing everything to a much better more coherent format that makes sense, yes this is broken code but I wanted to get a baseline of the new structure before implemented
