+++ loremaster/backend/package.json
@@ -10,7 +10,6 @@
   "license": "MIT",
   "dependencies": {
     "bcrypt": "^3.0.7",
-    "cloudinary": "^1.19.0",
     "express": "^4.17.1",
     "express-graphql": "^0.9.0",
     "graphql": "^14.5.4",
