+++ loremaster/backend/src/extraFunctions/generateId.js
@@ -2,7 +2,7 @@
 const { performance } = require('perf_hooks');
 
 module.exports = {
-    GenerateUuid: () => { // Public Domain/MIT
+    GenerateId: () => { // Public Domain/MIT
 
         var d = new Date().getTime(); //Timestamp
         var d2 = (performance && performance.now && (performance.now()*1000)) || 0;//Time in microseconds since page-load or 0 if unsupported
