+++ loremaster/backend/src/repositories/accountRepo.js
@@ -1,0 +1,10 @@
+const { Insert, Select } = require('../sql/sqlFunctions');
+
+module.exports = {
+    GetAccountWhere: () => {
+
+    },
+    CreateAccount: () => {
+        
+    }
+}
