+++ loremaster/backend/src/repositories/mapRepo.js
@@ -1,0 +1,10 @@
+const { Insert, Select } = require('../sql/sqlFunctions');
+
+module.exports = {
+    CreateAccount: () => {
+        
+    },
+    GetAccountWhere: () => {
+
+    }
+}
